Addressing issue with inconsistent imaris file structure.
Possibly a bug in the imaris software. When a channel was removed, the
group containing the channel data was removed, but the DataSetInfo
group associated with the channel was not removed. Consequently, the
file was in an "undefined state". When adding channels to a file we
expected that we would need to create the relevant DataSetInfo group
for each added channel. We now check if the group exists, even though
it shouldn't, and don't attempt to create it if it is already there.