Adding scroll area to the progress bar portion of the application.
When a large number of files was processed at the same time, the GUI
couldn't display all of the progress bars due to screen size. The
progress bars were updated but this was not visible to the user which
created the impression that the application was hanging/stuck. All of
these UI elements are now inside a scroll area so that the user can
see all the progress bars.