Improving support for color table.
The color table information was previously only stored in an
attribute. The problem is that HDF5 attributes have a size limit of
64K bytes. For large color tables imaris stores them in a
dataset. This commit adds that feature. Unfortunately, it isn't
complete as Imaris isn't able to read the dataset the way we write
it. Can't figure out how to get it working with h5py. For now the data
is there, can be exported using the XTExportChannelSettings extension
and then import the color tables back in using the imaris functionality.