Fixes two bugs, writing the image extent and image histogram.
In a previous commit which changed the format of writing the image
extend attributes (from NULL_PAD to NULL_TERM), also changed the
extend to write only three digits after the decimal point. That was
just a mistake.

Second error was in writing the image histogram for
pixel types that are not uint8. In this case there are two histogram
datasets, one named Histogram which has 256 bins and the other named
Histogram1024. A copy paste error named both of them as Histogram. As
all test images were uint8, this error was not detected. Adding uint16
image to the test suite addressed this oversight.