Adding check for image pixel type before starting registration flow.
We expect all images to have the same pixel type (uint8, uint16 or
float32). Previously, this was only checked during resampling after the
registration was performed and it threw an exception. The exception was
displayed to the user but was not understandable by non-computational
users. This modification explicitly checks pixel types before starting the
registration process and displays an error message which is
understandable to everyone.